Move filter config to another file
